REFACTOR the MQTT functionality out of the main file
