+++ CHANGES_NEXT_RELEASE
@@ -1,0 +1,1 @@
+- Modify the agent to allow multiple transport protocols (#109)
