+++ lib/iotaUtils.js
@@ -1,0 +1,79 @@
+/*
+ * Copyright 2016 Telefonica Investigación y Desarrollo, S.A.U
+ *
+ * This file is part of iotagent-json
+ *
+ * iotagent-json is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * iotagent-json is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public
+ * License along with iotagent-json.
+ * If not, seehttp://www.gnu.org/licenses/.
+ *
+ * For those usages not covered by the GNU Affero General Public License
+ * please contact with::[contacto@tid.es]
+ */
+
+var iotAgentLib = require('iotagent-node-lib'),
+    logger = require('logops'),
+    errors = require('./errors'),
+    constants = require('./constants'),
+    context = {
+        op: 'IoTAgentMQTT.Utils'
+    },
+    config = require('./configService');
+
+/**
+ * Get the API Key for the selected service if there is any, or the default API Key if a specific one does not exist.
+ *
+ * @param {String} service          Name of the service whose API Key we are retrieving.
+ * @param {String} subservice       Name of the subservice whose API Key we are retrieving.
+ */
+function getEffectiveApiKey(service, subservice, callback) {
+    logger.debug(context, 'Getting effective API Key');
+
+    iotAgentLib.findConfiguration(service, subservice, function(error, group) {
+        if (group) {
+            logger.debug(context, 'Using found group: %j', group);
+            callback(null, group.apikey);
+        } else if (config.getConfig().mqtt.defaultKey) {
+            logger.debug(context, 'Using default API Key: %s', config.getConfig().mqtt.defaultKey);
+            callback(null, config.getConfig().mqtt.defaultKey);
+        } else {
+            logger.ierror(context, 'Could not find any API Key information for device.');
+            callback(new errors.GroupNotFound(service, subservice));
+        }
+    });
+}
+
+/**
+ * Find the attribute given by its name between all the active attributes of the given device, returning its type, or
+ * null otherwise.
+ *
+ * @param {String} attribute        Name of the attribute to find.
+ * @param {Object} device           Device object containing all the information about a device.
+ * @return {String}                 String identifier of the attribute type.
+ */
+function guessType(attribute, device) {
+    for (var i = 0; i < device.active.length; i++) {
+        if (device.active[i].name === attribute) {
+            return device.active[i].type;
+        }
+    }
+
+    if (attribute === constants.TIMESTAMP_ATTRIBUTE) {
+        return constants.TIMESTAMP_TYPE;
+    } else {
+        return constants.DEFAULT_ATTRIBUTE_TYPE;
+    }
+}
+
+exports.guessType = guessType;
+exports.getEffectiveApiKey = getEffectiveApiKey;
